tf2_django_full_stack_app/requirements.txt
django
pillow
tensorflow